# CrawlerAPI/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Crawler API
  version: 1.0.0
  description: |
    OpenAPI specification for the Crawler API. Generated from Spring controller mappings
    under `src/main/java/com/crawlerApi/api`. Many endpoints require an authenticated account
    context (JWT) and return domain-specific DTOs for domains, audits, pages, elements, and test records.
servers:
  - url: http://localhost:9080
    description: Local default server port from application.properties
security:
  - bearerAuth: []
...
paths:
...
  /v1/auditor/start:
    post:
      tags: [Auditor]
      summary: Start audit
      description: Start an audit on the provided URL based on the audit level (page or domain).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuditStartRequest' }
      responses:
        '200': { description: Audit started, content: { application/json: { schema: { $ref: '#/components/schemas/AuditRecordDto' } } } }
        '400': { description: Invalid audit request (missing url or level) }
        '401': { description: Authentication required }
        '500': { description: Internal server error }
  /v1/auditor:
    get:
      tags: [Auditor]
      summary: Get page by URL
      description: Retrieve page information for the given URL.
      parameters:
        - in: query
          name: url
          required: true
          schema: { type: string, format: uri }
          description: Page URL
      responses:
        '200': { description: Page returned, content: { application/json: { schema: { $ref: '#/components/schemas/SimplePage' } } } }
        '400': { description: Invalid URL }
        '401': { description: Authentication required }
        '404': { description: Page not found }
        '500': { description: Internal server error }
  /v1/auditor/{page_key}/insights:
    get:
      tags: [Auditor]
      summary: Get page insights
      description: Retrieve performance insights for the given page.
      parameters:
        - in: path
          name: page_key
          required: true
          schema: { type: string }
          description: Key of the page
      responses:
        '200': { description: Insights returned, content: { application/json: { schema: { $ref: '#/components/schemas/PerformanceInsight' } } } }
        '401': { description: Authentication required }
        '403': { description: Insufficient permissions }
        '404': { description: Page not found }
...
  /v1/ide-test-export:
    put:
      tags: [IDE Test Export]
      summary: Update IDE test export record
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Object' } } }
      responses:
        '200': { description: Record updated }
    post:
      tags: [IDE Test Export]
      summary: Create IDE test export record
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Object' } } }
      responses:
        '200': { description: Record created }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Object:
      type: object
      additionalProperties: true
...
    AuditStartRequest:
      type: object
      required: [url, level]
      description: Request body for starting an audit
      properties:
        url:
          type: string
          format: uri
          description: URL to audit (page or domain root)
        level:
          type: string
          enum: [PAGE, DOMAIN]
          description: Audit level (single page or full domain)
    AuditRecordDto:
      type: object
      description: Audit record summary DTO
      additionalProperties: true
...
    PerformanceInsight:
      type: object
      description: Performance insights for a page
      additionalProperties: true
    SimplePage:
      type: object
      description: Simplified page representation
      properties:
        url: { type: string, format: uri }
        viewportScreenshotUrl: { type: string }
        fullPageScreenshotUrl: { type: string }
        fullPageWidth: { type: integer }
        fullPageHeight: { type: integer }
        src: { type: string }
        key: { type: string }
        id: { type: integer, format: int64 }
      additionalProperties: true
...